+++ src/Twig/Component/Icon.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Codeschubser\Bundle\TwigComponents\Twig\Component;
+namespace Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component;
 
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-#[AsTwigComponent(name: 'Icon', template: '@CodeschubserTwigComponents/components/Icon.html.twig', exposePublicProps: false)]
+#[AsTwigComponent(name: 'Icon', template: '@BootstrapTwigComponents/components/Icon.html.twig', exposePublicProps: false)]
 final class Icon
 {
     public string $name;