+++ tests/Integration/Twig/Components/ButtonTest.php
@@ -2,26 +2,29 @@
 
 declare(strict_types=1);
 
-namespace Codeschubser\Bundle\TwigComponents\Tests\Integration\Twig\Components;
+namespace Codeschubser\Bundle\BootstrapTwigComponentsBundle\Tests\Integration\Twig\Components;
 
-use Codeschubser\Bundle\TwigComponents\Tests\Common\AbstractComponentsTestCase;
-use Codeschubser\Bundle\TwigComponents\Twig\Component\Button;
-use Codeschubser\Bundle\TwigComponents\Twig\Component\Icon;
-use Codeschubser\Bundle\TwigComponents\Twig\Component\Option\ButtonType;
-use Codeschubser\Bundle\TwigComponents\Twig\Component\Option\Variant;
-use Codeschubser\Bundle\TwigComponents\Twig\Component\Option\VariantInterface;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Tests\Common\AbstractComponentsTestCase;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component\Button;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component\Icon;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component\Option\ButtonType;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component\Option\Variant;
+use Codeschubser\Bundle\BootstrapTwigComponentsBundle\Twig\Component\Option\VariantInterface;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\Attributes\UsesClass;
 
 #[CoversClass(Button::class)]
-#[CoversClass(Variant::class)]
-#[CoversClass(ButtonType::class)]
-#[CoversClass(Icon::class)]
+#[UsesClass(Variant::class)]
+#[UsesClass(ButtonType::class)]
+#[UsesClass(Icon::class)]
+#[Group('button')]
 final class ButtonTest extends AbstractComponentsTestCase
 {
     public function testComponentMount(): void
     {
         $component = $this->mountTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'danger',
             ],
@@ -37,7 +40,7 @@ final class ButtonTest extends AbstractComponentsTestCase
         $this->expectException(\InvalidArgumentException::class);
 
         $this->mountTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'foobar',
             ],
@@ -60,7 +63,7 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentWithIconMount(): void
     {
         $component = $this->mountTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'danger',
                 'icon' => 'bi bi-info-circle',
@@ -74,7 +77,7 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'success',
             ],
@@ -87,19 +90,19 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentWithOutlineVariantRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'outline-success',
             ],
         );
 
-        $this->assertCount(1, $component = $rendered->crawler()->filter('.btn.btn-outline-success'));
+        $this->assertCount(1, $rendered->crawler()->filter('.btn.btn-outline-success'));
     }
 
     public function testComponentAsSubmitRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'success',
                 'type' => 'submit',
@@ -113,7 +116,7 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentAsResetRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'secondary',
                 'type' => ButtonType::RESET,
@@ -127,7 +130,7 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentWithIconRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'warning',
                 'icon' => 'bi-exclamation-triangle-fill',
@@ -140,7 +143,7 @@ final class ButtonTest extends AbstractComponentsTestCase
     public function testComponentWithAdditionalParamsRenders(): void
     {
         $rendered = $this->renderTwigComponent(
-            name: 'Button',
+            name: Button::class,
             data: [
                 'variant' => 'warning',
                 'class' => 'btn-lg',
@@ -149,5 +152,17 @@ final class ButtonTest extends AbstractComponentsTestCase
 
         $this->assertCount(1, $component = $rendered->crawler()->filter('.btn.btn-warning'));
         $this->assertCount(1, $component->filter('.btn.btn-lg'));
+    }
+
+    public function testComponentAsCloseButton(): void
+    {
+        $rendered = $this->renderTwigComponent(
+            name: Button::class,
+            data: [
+                'variant' => 'close',
+            ],
+        );
+
+        $this->assertCount(1, $rendered->crawler()->filter('.btn-close'));
     }
 }